// frontend/src/app/services/models/page-response-borrowed-product-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { BorrowedProductResponse } from '../models/borrowed-product-response';
export interface PageResponseBorrowedProductResponse {
  content?: Array<BorrowedProductResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}